Fix controller misplacements and naming
Split ProfileController: auth routes (sign-up, sign-in, email
verification) move to accounts/AuthPagesController. Profile
view/edit stays in user/ProfileController.

Rename HomeController → PublicController to match its domain.
Merge DebugController + EditorTestController into single file.

Add future path rename notes to restructure plan.
