+++ convergence_games/apps/frontend/accounts/controllers/__init__.py
@@ -1,4 +1,5 @@
+from ._auth_pages import AuthPagesController
 from ._email_auth import EmailAuthController
 from ._oauth import OAuthController
 
-__all__ = ["EmailAuthController", "OAuthController"]
+__all__ = ["AuthPagesController", "EmailAuthController", "OAuthController"]
