+++ convergence_games/apps/frontend/debug/controllers/__init__.py
@@ -1,4 +1,3 @@
 from ._debug import DebugController
-from ._editor_test import EditorTestController
 
-__all__ = ["DebugController", "EditorTestController"]
+__all__ = ["DebugController"]
