+++ convergence_games/apps/frontend/debug/controllers/_debug.py
@@ -1,13 +1,14 @@
 from litestar import Controller, get
 
-from convergence_games.lib.response_type import HTMXBlockTemplate
+from convergence_games.lib.request_type import Request
+from convergence_games.lib.response_type import HTMXBlockTemplate, Template
 
 
 class DebugController(Controller):
     path = "/debug"
 
     @get(path="/email-template")
-    async def index(self) -> HTMXBlockTemplate:
+    async def get_email_template(self) -> HTMXBlockTemplate:
         return HTMXBlockTemplate(
             template_name="emails/sign_in_code.html.jinja",
             context={
@@ -16,3 +17,7 @@ class DebugController(Controller):
                 "expires_at": "2023-10-01T00:00:00Z",
             },
         )
+
+    @get(path="/editor-test")
+    async def get_editor_test(self, request: Request) -> Template:
+        return HTMXBlockTemplate(template_name="pages/editor_test.html.jinja", block_name=request.htmx.target)
