+++ convergence_games/apps/frontend/public/controllers/_public.py
@@ -4,7 +4,7 @@ from convergence_games.lib.request_type import Request
 from convergence_games.lib.response_type import HTMXBlockTemplate, Template
 
 
-class HomeController(Controller):
+class PublicController(Controller):
     @get(path="/")
     async def get_home(self, request: Request) -> Template:
         return HTMXBlockTemplate(template_name="pages/home.html.jinja", block_name=request.htmx.target)
